+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Redesign "Metadata" tab
+- Updated the algorithm for checking the hash of the head
+### Fixed
+- Require initial choice of radio buttons in transform options form
+- Time delta form is not disabled initially in ingest schedule
+- Improved setting form labels - auto disable when control is disabled
 
 ## [0.55.0] - 2025-08-20
 ### Added